Replace `fetchApi` with `useFetch`
